requirements.txt
apache-airflow[celery]==3.1.6
apache-airflow-providers-apache-spark
pyspark==3.5.0
pyiceberg[pyarrow,pandas,s3fs]
minio
pandas
kaggle